Clean up files, name variables, other refactoring
